django/requirements.txt
asgiref
CacheControl
cachetools
certifi
cffi
chardet
charset-normalizer
cryptography
defusedxml
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
djoser
firebase-admin
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
httplib2
idna
msgpack
oauthlib
pillow
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
PyJWT
pyparsing
python3-openid
reportlab
requests
requests-oauthlib
rsa
social-auth-app-django
social-auth-core
sqlparse
tzdata
uritemplate
urllib3
dj-database-url
gunicorn
pillow
reportlab
